+++ frontend/src/__generated/ve/components/Search/SavedSegmentQueryDisplay/styles.css.js
@@ -1,0 +1,9 @@
+// src/components/Search/SavedSegmentQueryDisplay/styles.css.ts
+var comboboxTag = "_1wt9z3q1";
+var comboboxTagBackground = "_1wt9z3q2";
+var comboboxTagsContainer = "_1wt9z3q0";
+export {
+  comboboxTag,
+  comboboxTagBackground,
+  comboboxTagsContainer
+};
