--- frontend/src/__generated/ve/pages/LogsPage/SavedSegmentQueryDisplay/styles.css.js
@@ -1,9 +1,0 @@
-// src/pages/LogsPage/SavedSegmentQueryDisplay/styles.css.ts
-var comboboxTag = "_7g52oo1";
-var comboboxTagBackground = "_7g52oo2";
-var comboboxTagsContainer = "_7g52oo0";
-export {
-  comboboxTag,
-  comboboxTagBackground,
-  comboboxTagsContainer
-};
