+++ frontend/src/components/Search/SavedSegmentQueryDisplay/styles.css.ts
